requirements.txt
fastapi
uvicorn
stripe
supabase
python-dotenv
